+++ packages/mobile/src/App.tsx
@@ -46,6 +46,8 @@ import { RootStackParamList } from './route.params'
 import ConnectionProcessScreen from './screens/ConnectionProcess/ConnectionProcess.screen'
 import { DuplicatedUsernameScreen } from './screens/DuplicatedUsername/DuplicatedUsername.screen'
 import { UnregisteredUsernameScreen } from './screens/UnregisteredUsername/UnregisteredUsername.screen'
+import UsernameTakenScreen from './screens/UsernameTaken/UsernameTaken.screen'
+import NewUsernameRequestedScreen from './screens/NewUsernameRequested/NewUsernameRequested.screen'
 import { PossibleImpersonationAttackScreen } from './screens/PossibleImpersonationAttack/PossibleImpersonationAttack.screen'
 
 LogBox.ignoreAllLogs()
@@ -104,6 +106,8 @@ function App(): JSX.Element {
                 <Screen component={ErrorScreen} name={ScreenNames.ErrorScreen} />
                 <Screen component={DuplicatedUsernameScreen} name={ScreenNames.DuplicatedUsernameScreen} />
                 <Screen component={UnregisteredUsernameScreen} name={ScreenNames.UnregisteredUsernameScreen} />
+                <Screen component={UsernameTakenScreen} name={ScreenNames.UsernameTakenScreen} />
+                <Screen component={NewUsernameRequestedScreen} name={ScreenNames.NewUsernameRequestedScreen} />
                 <Screen
                   component={PossibleImpersonationAttackScreen}
                   name={ScreenNames.PossibleImpersonationAttackScreen}