+++ packages/mobile/src/const/ScreenNames.enum.ts
@@ -14,5 +14,7 @@ export enum ScreenNames {
   ConnectionProcessScreen = 'ConnectionProcessScreen',
   DuplicatedUsernameScreen = 'DuplicatedUsernameScreen',
   UnregisteredUsernameScreen = 'UnregisteredUsernameScreen',
+  UsernameTakenScreen = 'UsernameTakenScreen',
+  NewUsernameRequestedScreen = 'NewUsernameRequestedScreen',
   PossibleImpersonationAttackScreen = 'PossibleImpersonationAttackScreen',
 }