+++ packages/mobile/src/store/navigation/redirection/redirection.saga.ts
@@ -3,7 +3,7 @@ import { initSelectors } from '../../init/init.selectors'
 import { navigationSelectors } from '../navigation.selectors'
 import { navigationActions } from '../navigation.slice'
 import { ScreenNames } from '../../../const/ScreenNames.enum'
-import { identity, users } from '@quiet/state-manager'
+import { identity, publicChannels, users } from '@quiet/state-manager'
 import { initActions } from '../../init/init.slice'
 
 export function* redirectionSaga(): Generator {
@@ -25,6 +25,16 @@ export function* redirectionSaga(): Generator {
     return
   }
 
+  const isUsernameTaken = yield* select(identity.selectors.usernameTaken)
+
+  if (isUsernameTaken) {
+    yield* put(
+      navigationActions.replaceScreen({
+        screen: ScreenNames.UsernameTakenScreen,
+      })
+    )
+  }
+
   const duplicateCerts = yield* select(users.selectors.duplicateCerts)
 
   if (duplicateCerts) {
@@ -37,9 +47,12 @@ export function* redirectionSaga(): Generator {
   }
 
   // If user belongs to a community, let him directly into the app
+  // Check while QA session!
+  const currentChannelDisplayableMessages = yield* select(publicChannels.selectors.currentChannelMessagesMergedBySender)
+  const areMessagesLoaded = Object.values(currentChannelDisplayableMessages).length > 0
   const communityMembership = yield* select(identity.selectors.communityMembership)
 
-  if (communityMembership) {
+  if (communityMembership && areMessagesLoaded) {
     yield* put(
       navigationActions.replaceScreen({
         screen: ScreenNames.ChannelListScreen,