+++ packages/state-manager/src/sagas/identity/identity.selectors.ts
@@ -3,6 +3,7 @@ import { createSelector } from '@reduxjs/toolkit'
 import { identityAdapter } from './identity.adapter'
 import { type CreatedSelectors, type StoreState } from '../store.types'
 import { communitiesSelectors, selectCommunities } from '../communities/communities.selectors'
+import { certificatesMapping } from '../users/users.selectors'
 
 const identitySlice: CreatedSelectors[StoreKeys.Identity] = (state: StoreState) => state[StoreKeys.Identity]
 
@@ -22,6 +23,10 @@ export const currentIdentity = createSelector(
 )
 
 export const communityMembership = createSelector(currentIdentity, identity => {
+  return Boolean(identity?.userCsr)
+})
+
+export const hasCertificate = createSelector(currentIdentity, identity => {
   return Boolean(identity?.userCertificate)
 })
 
@@ -37,6 +42,21 @@ export const csr = createSelector(communitiesSelectors.currentCommunityId, selec
   return identities[id]?.userCsr
 })
 
+export const usernameTaken = createSelector(currentIdentity, certificatesMapping, (identity, certs) => {
+  const userCertificate = identity?.userCertificate
+  if (userCertificate) return false
+
+  const username = identity?.nickname
+  if (!username) return false
+
+  const allUsersSet = new Set(Object.values(certs).map(user => user.username))
+  if (allUsersSet.has(username)) {
+    return true
+  }
+
+  return false
+})
+
 export const identitySelectors = {
   selectById,
   selectEntities,
@@ -45,4 +65,6 @@ export const identitySelectors = {
   joinedCommunities,
   joinTimestamp,
   csr,
+  usernameTaken,
+  hasCertificate,
 }