+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 
 [unreleased]
 
+* UI layer for taken usernames for desktop and mobile
+
+* Change nickname for taken username
+
+* Map messages sent before changing username
+
+* Update registrar service to match new registration flow.
+
 * Add possible impersonation attack UI for desktop and mobile
 
 * Fix truncated long messages in channelInput component